requirements.txt
Django==1.10.5
django-cors-headers==2.0.2
django-extensions==1.7.6
djangorestframework==3.5.4
future==0.16.0
line-bot-sdk==1.1.0
olefile==0.44
requests==2.13.0
Pillow==4.0.0
pyhamcrest==1.9.0
pytz==2016.10
jupyter==1.0.0
jieba==0.38
celery==4.0.2
django-celery-beat==1.0.1
